requirements.txt
RPI.GPIO
adafruit-blinka
rpi_ws281x
adafruit-circuitpython-neopixel
azure-iot-device
iotc
python-dotenv
